Oops

+++ src/main/java/cn/nukkit/network/protocol/BossEventPacket.java
@@ -7,7 +7,7 @@ public class BossEventPacket extends DataPacket {
 
     public static final byte NETWORK_ID = ProtocolInfo.BOSS_EVENT_PACKET;
 
-    public int eid;
+    public long eid;
     public int type;
 
     @Override
